// src/scripts/config.d.ts
export declare const ROAD_WIDTH = 500;
export declare const VEHICLE_WIDTH = 50;
export declare const VEHICLE_HEIGHT = 100;
export declare const MARKING_WIDTH = 10;
export declare const MARKING_HEIGHT: number;
export declare const TRAFFIC_LANES = 6;
export declare const COMPETITORS_COUNT = 6;
export declare const EXPLOSIVE_WIDTH = 75;
declare type Settings = {
    start: boolean;
    score: number;
    flowSpeed: number;
    traffic: number;
    maxCompetitorsViews: number;
    paused: boolean;
    audio: boolean;
    backgroundVideo: boolean;
    x: number;
    y: number;
};
export declare const SETTINGS: Settings;
export declare const setStyles: () => void;
export {};
//# sourceMappingURL=config.d.ts.map